# agentauth/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentAuth API
  version: 0.7.0
  description: |
    **AgentAuth** is a production-ready authentication and identity service for AI agents.

    ## Features
    - JWT-based authentication
    - Rate limiting (100 req/15min general, 10 req/15min auth)
    - Webhooks for events
    - Activity logging
    - Permission-based access control
    - **Persona System** — define and manage an agent's digital soul (v0.7.0)
    - **ZKP Anonymous Verification** — prove identity without revealing credentials (v0.7.0)
    - **Anti-Drift Vault** — detect behavioral drift and auto-revoke rogue agents (v0.7.0)

    ## Authentication
    Most endpoints require an API key in the `X-Api-Key` header:
    ```
    X-Api-Key: ag_sk_your_api_key_here
    ```

    ## Rate Limits
    - **General endpoints**: 100 requests per 15 minutes
    - **Authentication endpoints**: 10 requests per 15 minutes

    Rate limit info is included in response headers:
    - `X-RateLimit-Limit`: Max requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time until limit resets (Unix timestamp)

    ## Base URL
    - **Production**: `https://api.agentauth.dev`
    - **Local**: `http://localhost:3000`

  contact:
    name: AgentAuth Support
    email: umytbaynazarow754@gmail.com
    url: https://agentauth.dev
  license:
    name: MIT
    url: https://github.com/umyt-dev/Agent-Identity/blob/main/LICENSE
servers:
  - url: https://api.agentauth.dev
...
paths:
...
  /v1/agents/verify:
    post:
      summary: Verify agent credentials
      description: |
        Verify agent credentials (agent_id + API key) and receive agent details.

        **Rate limit:** 10 requests per 15 minutes
      operationId: verifyAgent
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentVerificationRequest'
            example:
              agent_id: ag_1a2b3c4d5e6f
              api_key: ag_sk_1a2b3c4d5e6f7g8h9i0j
      responses:
        '200':
          description: Agent verified successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVerificationResponse'
              example:
                valid: true
                agent_id: ag_1a2b3c4d5e6f
                name: CustomerSupportAgent
                tier: free
                permissions:
                  - read:messages
                  - write:responses
                status: active
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/zkp/register-commitment:
    post:
      summary: Register ZKP commitment
      description: |
        Generate and store a SHA-256 commitment for anonymous verification.
        Requires valid agent credentials (agent_id + api_key in body).

        The returned `salt` is only shown once. Store it securely.
      operationId: registerCommitment
      tags:
        - ZKP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitmentRegistrationRequest'
            example:
              agent_id: ag_1a2b3c4d5e6f
              api_key: ag_sk_1a2b3c4d5e6f7g8h9i0j
              expires_in: 86400
      responses:
        '201':
          description: Commitment registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitmentRegistrationResponse'
              example:
                commitment: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
                salt: f1e2d3c4b5a6f1e2d3c4b5a6f1e2d3c4b5a6f1e2d3c4b5a6f1e2d3c4b5a6f1e2
                expires_at: '2026-02-02T12:00:00.000Z'
                message: Store the salt securely — it will not be shown again.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/zkp/verify-anonymous:
    post:
      summary: Verify anonymously
      description: |
        Verify a commitment without revealing agent identity.

        **Hash mode** (`?mode=hash`): Provide `preimage_hash` — the SHA-256 of
        `agentId:apiKey:salt`. Fast, no cryptographic proof required.

        **ZKP mode** (`?mode=zkp`, default): Provide a Groth16 proof and
        publicSignals. The first public signal must equal the commitment.
      operationId: verifyAnonymous
      tags:
        - ZKP
      parameters:
        - name: mode
          in: query
          description: Verification mode
          required: false
          schema:
            type: string
            enum: [zkp, hash]
            default: zkp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ZKPVerificationRequest'
                - $ref: '#/components/schemas/HashVerificationRequest'
            examples:
              zkp_mode:
                summary: ZKP mode with Groth16 proof
                value:
                  commitment: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
                  proof:
                    pi_a: ['0x1234...', '0x5678...', '1']
                    pi_b: [['0xabcd...', '0xef01...'], ['0x2345...', '0x6789...']]
                    pi_c: ['0xaaaa...', '0xbbbb...', '1']
                  publicSignals:
                    - a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
              hash_mode:
                summary: Hash mode with preimage
                value:
                  commitment: a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2
                  preimage_hash: abcd1234abcd1234abcd1234abcd1234abcd1234abcd1234abcd1234abcd1234
      responses:
        '200':
          description: Verification result
          headers:
            Cache-Control:
              description: no-store to prevent caching of verification results
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnonymousVerificationResponse'
              examples:
                valid:
                  value:
                    valid: true
                    reason: Commitment verified via hash mode
                    permissions:
                      - read
                      - write
                    tier: pro
                invalid:
                  value:
                    valid: false
                    reason: Preimage hash does not match commitment
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/zkp/commitment/{commitment}:
    delete:
      summary: Revoke commitment
      description: Mark a commitment as revoked so it can no longer be used for verification.
      operationId: revokeCommitment
      tags:
        - ZKP
      parameters:
        - name: commitment
          in: path
          description: The commitment hex string to revoke
          required: true
          schema:
            type: string
            pattern: '^[0-9a-f]{64,256}$'
      responses:
        '200':
          description: Commitment revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  commitment:
                    type: string
                  status:
                    type: string
              example:
                message: Commitment revoked
                commitment: a1b2c3d4...
                status: revoked
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/zkp/active-count:
    get:
      summary: Count active commitments
      description: Return the number of currently active (non-revoked, non-expired) commitments.
      operationId: getActiveCommitmentCount
      tags:
        - ZKP
      responses:
        '200':
          description: Active commitment count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
              example:
                count: 42
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─── Drift (Anti-Drift Vault) ───────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Agent API key
...
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
      example: 10
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
      example: 7
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1738329600
  schemas:
...
    AgentVerificationRequest:
      type: object
      required: [agent_id, api_key]
      properties:
        agent_id:
          type: string
        api_key:
          type: string
    AgentVerificationResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        agent_id:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        error:
          type: string
...
    CommitmentRegistrationRequest:
      type: object
      required: [agent_id, api_key]
      properties:
        agent_id:
          type: string
        api_key:
          type: string
        expires_in:
          type: integer
          minimum: 1
          description: TTL in seconds (optional)
    CommitmentRegistrationResponse:
      type: object
      properties:
        commitment:
          type: string
          description: 64-char hex SHA-256 commitment
        salt:
          type: string
          description: 64-char hex random salt (shown once)
        expires_at:
          type: string
          format: date-time
          nullable: true
        message:
          type: string
    Groth16Proof:
      type: object
      required: [pi_a, pi_b, pi_c]
      description: Groth16 zero-knowledge proof
      properties:
        pi_a:
          type: array
          items:
            type: string
        pi_b:
          type: array
          items:
            type: array
            items:
              type: string
        pi_c:
          type: array
          items:
            type: string
        protocol:
          type: string
          enum: [groth16]
        curve:
          type: string
          enum: [bn128, bls12381]
    ZKPVerificationRequest:
      type: object
      required: [commitment, proof, publicSignals]
      properties:
        commitment:
          type: string
          pattern: '^[0-9a-f]{64,256}$'
        proof:
          $ref: '#/components/schemas/Groth16Proof'
        publicSignals:
          type: array
          items:
            type: string
          description: First element must equal the commitment
    HashVerificationRequest:
      type: object
      required: [commitment, preimage_hash]
      properties:
        commitment:
          type: string
          pattern: '^[0-9a-f]{64,256}$'
        preimage_hash:
          type: string
          description: SHA-256 of "agentId:apiKey:salt"
    AnonymousVerificationResponse:
      type: object
      required: [valid, reason]
      properties:
        valid:
          type: boolean
        reason:
          type: string
        permissions:
          type: array
          items:
            type: string
          description: Agent permissions (only if valid)
        tier:
          type: string
          description: Agent tier (only if valid)

    # ─── Drift Schemas ───────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            message: Invalid input data
            details:
              - field: email
                message: Must be a valid email address
    Unauthorized:
      description: Unauthorized - invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid agent credentials
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Resource not found
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate Limit Exceeded
            message: Too many requests. Please try again later.
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal Server Error
            message: An unexpected error occurred